# backendnest/swagger.yaml
openapi: 3.0.0
info:
  title: API de Gestão Financeira Pessoal
  description: |
    API REST para autenticação e gestão de contas, categorias, transações,
    transferências, dívidas, metas e alertas.

    ## Convenção de idioma
    Os nomes dos campos seguem **camelCase em português**, pois as entidades
    do banco de dados estão em português (ex. `nome`, `senha`, `saldoInicial`).
    Alguns campos de domínio financeiro estão em espanhol (ex. `montoTotal`,
    `fechaVencimiento`, `diasAnticipacion`).

    ## Autenticação
    Todos os endpoints (exceto `/auth/register` e `/auth/login`) requerem um
    token JWT no header `Authorization: Bearer <token>`.
    O token expira em **1 hora**.

    ## Códigos de status
    - `200` — Operação bem-sucedida (GET, PATCH, DELETE)
    - `201` — Recurso criado (POST)
    - `400` — Validação de campos falhou
    - `401` — Token ausente, expirado ou inválido
    - `404` — Recurso não encontrado ou não pertence ao usuário
    - `409` — Conflito (ex. e-mail duplicado)
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Servidor local de desenvolvimento
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido em /auth/login. Expira em 1 hora.
  schemas:
...
    CreateTransacaoRequest:
      type: object
      required: [contaId, categoriaId, tipo, valor, data]
      properties:
        contaId:
          type: string
          format: uuid
          description: ID da conta associada (deve existir e pertencer ao usuário).
        categoriaId:
          type: string
          format: uuid
          description: ID da categoria associada (deve existir e pertencer ao usuário).
        tipo:
          type: string
          enum: [receita, despesa]
          description: |
            Tipo de transação:
            - `receita` — Receita (entrada)
            - `despesa` — Despesa (saída)
        valor:
          type: number
          format: double
          description: Valor monetário (decimal 12,2). Sempre positivo.
          example: 50.00
        data:
          type: string
          format: date
          description: Data da transação (formato YYYY-MM-DD).
          example: "2026-03-30"
        descricao:
          type: string
          nullable: true
          description: Descrição opcional da transação.
        ehAjuste:
          type: boolean
          default: false
          description: Se é um ajuste manual de saldo (não uma transação real).
    UpdateTransacaoRequest:
      type: object
      description: Apenas os campos enviados serão atualizados. Não é permitido alterar `tipo` nem `ehAjuste`.
      properties:
        contaId:
          type: string
          format: uuid
          description: Nova conta associada.
        categoriaId:
          type: string
          format: uuid
          description: Nova categoria associada.
        valor:
          type: number
          format: double
          description: Novo valor monetário.
        data:
          type: string
          format: date
          description: Nova data.
        descricao:
          type: string
          description: Nova descrição.
    Transacao:
      type: object
      description: Entidade completa de uma transação.
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único.
        usuarioId:
          type: string
          format: uuid
          description: ID do usuário dono.
        contaId:
          type: string
          format: uuid
          description: ID da conta associada.
        categoriaId:
          type: string
          format: uuid
          description: ID da categoria associada.
        tipo:
          type: string
          enum: [receita, despesa]
          description: Tipo (receita ou despesa).
        valor:
          type: number
          format: double
          description: Valor monetário.
        data:
          type: string
          format: date
          description: Data da transação.
        descricao:
          type: string
          nullable: true
          description: Descrição opcional.
        ehAjuste:
          type: boolean
          description: Se é um ajuste manual de saldo.
        createdAt:
          type: string
          format: date-time
          description: Data de criação do registro.
        updatedAt:
          type: string
          format: date-time
          description: Data da última atualização.

    # ══════════════════════════════════════
    #  TRANSFERENCIAS
    # ══════════════════════════════════════
...
    CreateDividaRequest:
      type: object
      required: [nome, montoTotal, fechaInicio, fechaVencimiento]
      properties:
        nome:
          type: string
          description: Nome descritivo da dívida.
          example: Empréstimo pessoal
        montoTotal:
          type: number
          format: double
          description: Valor total da dívida (decimal 12,2).
          example: 5000.00
        fechaInicio:
          type: string
          format: date
          description: Data em que a dívida foi contraída.
        fechaVencimiento:
          type: string
          format: date
          description: Data limite para pagamento total.
        contaId:
          type: string
          format: uuid
          nullable: true
          description: Conta associada (opcional). Se informada, deve existir e pertencer ao usuário.
        tasaInteres:
          type: number
          format: double
          nullable: true
          description: Taxa de juros (decimal 5,2). Ex. 2.50 = 2,5%.
        cuotaMensual:
          type: number
          format: double
          nullable: true
          description: Valor da parcela mensal (se aplicável).
        proximoVencimiento:
          type: string
          format: date
          nullable: true
          description: Data do próximo vencimento de parcela.
        periodicidade:
          type: string
          enum: [semanal, quinzenal, mensal, anual]
          nullable: true
          description: |
            Periodicidade de pagamento:
            - `semanal` — Semanal
            - `quinzenal` — Quinzenal
            - `mensal` — Mensal
            - `anual` — Anual
    UpdateDividaRequest:
      type: object
      description: Apenas os campos enviados serão atualizados. Não é permitido alterar `montoTotal` nem `fechaInicio`.
      properties:
        nome:
          type: string
          description: Novo nome.
        tasaInteres:
          type: number
          format: double
          description: Nova taxa de juros.
        cuotaMensual:
          type: number
          format: double
          description: Novo valor da parcela mensal.
        fechaVencimiento:
          type: string
          format: date
          description: Nova data de vencimento total.
        proximoVencimiento:
          type: string
          format: date
          description: Nova data do próximo vencimento.
        periodicidade:
          type: string
          enum: [semanal, quinzenal, mensal, anual]
          description: Nova periodicidade.
        ativa:
          type: boolean
          description: Estado ativo/inativo.
    Divida:
      type: object
      description: Entidade completa de uma dívida.
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único.
        usuarioId:
          type: string
          format: uuid
          description: ID do usuário dono.
        contaId:
          type: string
          format: uuid
          nullable: true
          description: Conta associada (pode ser null).
        nome:
          type: string
          description: Nome da dívida.
        montoTotal:
          type: number
          format: double
          description: Valor total da dívida.
        tasaInteres:
          type: number
          nullable: true
          description: Taxa de juros.
        cuotaMensual:
          type: number
          nullable: true
          description: Valor da parcela mensal.
        fechaInicio:
          type: string
          format: date
          description: Data de início.
        fechaVencimiento:
          type: string
          format: date
          description: Data de vencimento total.
        proximoVencimiento:
          type: string
          format: date
          nullable: true
          description: Data do próximo vencimento de parcela.
        periodicidade:
          type: string
          enum: [semanal, quinzenal, mensal, anual]
          nullable: true
          description: Periodicidade de pagamento.
        ativa:
          type: boolean
          description: Se a dívida está ativa.
        createdAt:
          type: string
          format: date-time
          description: Data de criação.

    # ══════════════════════════════════════
    #  PAGOS DIVIDA
    # ══════════════════════════════════════
    CreatePagoDividaRequest:
      type: object
      required: [dividaId, contaId, categoriaId, valor, data]
      properties:
        dividaId:
          type: string
          format: uuid
          description: ID da dívida à qual o pagamento se aplica.
        contaId:
          type: string
          format: uuid
          description: ID da conta de onde o pagamento é feito.
        categoriaId:
          type: string
          format: uuid
          description: ID da categoria para a transação de despesa criada automaticamente.
        valor:
          type: number
          format: double
          description: Valor do pagamento (decimal 12,2).
          example: 500.00
        data:
          type: string
          format: date
          description: Data do pagamento.
        descricao:
          type: string
          nullable: true
          description: Descrição opcional do pagamento.
    PagoDivida:
      type: object
      description: |
        Entidade de pagamento de dívida. Ao criar um pagamento, é gerada automaticamente
        uma transação do tipo `despesa` na mesma operação atômica.
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único do pagamento.
        usuarioId:
          type: string
          format: uuid
          description: ID do usuário dono.
        dividaId:
          type: string
          format: uuid
          description: ID da dívida associada.
        contaId:
          type: string
          format: uuid
          description: ID da conta de onde foi pago.
        transacaoId:
          type: string
          format: uuid
          description: ID da transação de despesa gerada automaticamente (relação 1-para-1).
        valor:
          type: number
          format: double
          description: Valor do pagamento.
        data:
          type: string
          format: date
          description: Data do pagamento.
        descricao:
          type: string
          nullable: true
          description: Descrição opcional.
        createdAt:
          type: string
          format: date-time
          description: Data de criação.

    # ══════════════════════════════════════
    #  METAS
    # ══════════════════════════════════════
    CreateMetaRequest:
      type: object
      required: [nome, tipo, montoObjetivo, fechaLimite]
      properties:
        nome:
          type: string
          description: Nome descritivo da meta.
          example: Viagem de férias
        tipo:
          type: string
          enum: [economia, reducao_divida]
          description: |
            Tipo de meta:
            - `economia` — Meta de economia
            - `reducao_divida` — Meta de redução de dívida
        montoObjetivo:
          type: number
          format: double
          description: Valor objetivo a alcançar (decimal 12,2).
          example: 3000.00
        fechaLimite:
          type: string
          format: date
          description: Data limite para cumprir a meta.
        contaId:
          type: string
          format: uuid
          nullable: true
          description: Conta associada (opcional, para metas de economia).
        dividaId:
          type: string
          format: uuid
          nullable: true
          description: Dívida associada (opcional, para metas de redução de dívida).
...
    Meta:
      type: object
      description: Entidade completa de uma meta financeira.
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único.
        usuarioId:
          type: string
          format: uuid
          description: ID do usuário dono.
        nome:
          type: string
          description: Nome da meta.
        tipo:
          type: string
          enum: [economia, reducao_divida]
          description: Tipo de meta.
        montoObjetivo:
          type: number
          format: double
          description: Valor objetivo.
        montoActual:
          type: number
          format: double
          description: Valor atual acumulado (padrão 0).
        fechaLimite:
          type: string
          format: date
          description: Data limite.
        contaId:
          type: string
          format: uuid
          nullable: true
          description: Conta associada.
        dividaId:
          type: string
          format: uuid
          nullable: true
          description: Dívida associada.
        ativa:
          type: boolean
          description: Se a meta está ativa.
        createdAt:
          type: string
          format: date-time
          description: Data de criação.

    # ══════════════════════════════════════
    #  ALERTAS
    # ══════════════════════════════════════
...
    Error400:
      type: object
      description: Erro de validação. O campo `message` pode ser string ou array de strings com os erros específicos.
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          description: Detalhe do erro de validação.
        error:
          type: string
          example: Bad Request
...
    Error404:
      type: object
      description: Recurso não encontrado ou não pertence ao usuário autenticado.
      properties:
        statusCode:
          type: integer
          example: 404
        message:
          type: string
          example: Conta não encontrada
...
  responses:
    BadRequest:
      description: Validação de campos falhou.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error400'
    Unauthorized:
      description: Token JWT ausente, expirado ou inválido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error401'
    NotFound:
      description: Recurso não encontrado ou não pertence ao usuário autenticado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error404'
...
paths:
...
  /transacoes:
    post:
      tags: [Transacoes]
      summary: Criar transação
      description: |
        Cria uma nova transação. Valida que `contaId` e `categoriaId`
        existam e pertençam ao usuário antes de criar.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransacaoRequest'
      responses:
        '201':
          description: Transação criada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transacao'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Conta ou categoria não encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
    get:
      tags: [Transacoes]
      summary: Listar transações
      description: |
        Retorna todas as transações do usuário autenticado,
        ordenadas por data **descendente** (mais recentes primeiro).
        Atualmente não suporta filtros por query params.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de transações ordenadas por data DESC.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transacao'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /transacoes/{id}:
    get:
      tags: [Transacoes]
      summary: Obter transação por ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID da transação.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detalhe da transação.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transacao'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Transacoes]
      summary: Atualizar transação
      description: Atualiza os campos enviados. Retorna a transação atualizada.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID da transação.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTransacaoRequest'
      responses:
        '200':
          description: Transação atualizada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transacao'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Transacoes]
      summary: Excluir transação
      description: Exclui a transação permanentemente (hard delete). Não retorna body.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID da transação.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Transação excluída. Sem body de resposta.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ════════════════════════════════════════
  #  TRANSFERENCIAS
  # ════════════════════════════════════════
...
  /dividas:
    post:
      tags: [Dividas]
      summary: Criar dívida
      description: |
        Cria uma nova dívida. Se `contaId` for informado, valida que a conta exista e pertença ao usuário.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDividaRequest'
      responses:
        '201':
          description: Dívida criada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Divida'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    get:
      tags: [Dividas]
      summary: Listar dívidas
      description: |
        Retorna todas as dívidas do usuário autenticado,
        ordenadas por `proximoVencimiento` **ascendente** (mais próximas de vencer primeiro).
        Atualmente não suporta filtros por query params.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de dívidas ordenadas por próximo vencimento ASC.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Divida'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /dividas/{id}:
    get:
      tags: [Dividas]
      summary: Obter dívida por ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID da dívida.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detalhe da dívida.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Divida'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Dividas]
      summary: Atualizar dívida
      description: Atualiza os campos enviados. Retorna a dívida atualizada.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID da dívida.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDividaRequest'
      responses:
        '200':
          description: Dívida atualizada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Divida'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /dividas/{id}/desativar:
    patch:
      tags: [Dividas]
      summary: Desativar dívida (soft delete)
      description: Marca a dívida como inativa. Não retorna body.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID da dívida.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Dívida desativada. Sem body de resposta.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ════════════════════════════════════════
  #  PAGOS DE DIVIDA
  # ════════════════════════════════════════
  /pagos-divida:
    post:
      tags: [Pagos de Divida]
      summary: Criar pagamento de dívida
      description: |
        Cria um pagamento de dívida como **operação atômica**:
        1. Cria uma transação do tipo `despesa` na conta indicada
        2. Cria o registro de pagamento vinculado a essa transação

        Se qualquer etapa falhar, ambas são revertidas (rollback).
        Valida que `dividaId`, `contaId` e `categoriaId` existam e pertençam ao usuário.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePagoDividaRequest'
      responses:
        '201':
          description: Pagamento criado com sucesso (junto com a transação associada).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagoDivida'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Dívida, conta ou categoria não encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
  /pagos-divida/divida/{dividaId}:
    get:
      tags: [Pagos de Divida]
      summary: Listar pagamentos de uma dívida
      description: Retorna todos os pagamentos de uma dívida específica, ordenados por data **descendente**.
      security:
        - BearerAuth: []
      parameters:
        - name: dividaId
          in: path
          required: true
          description: UUID da dívida.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lista de pagamentos da dívida ordenados por data DESC.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PagoDivida'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /pagos-divida/{id}:
    get:
      tags: [Pagos de Divida]
      summary: Obter pagamento por ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID do pagamento.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detalhe do pagamento.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagoDivida'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Pagos de Divida]
      summary: Excluir pagamento de dívida
      description: |
        Exclui o pagamento e sua transação associada como **operação atômica**.
        Ambos os registros são excluídos ou nenhum (rollback em caso de falha).
        Não retorna body.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID do pagamento.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Pagamento e transação excluídos. Sem body de resposta.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ════════════════════════════════════════
  #  METAS
  # ════════════════════════════════════════
  /metas:
    post:
      tags: [Metas]
      summary: Criar meta
      description: |
        Cria uma meta financeira. Se `contaId` ou `dividaId` for informado,
        valida que existam e pertençam ao usuário.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMetaRequest'
      responses:
        '201':
          description: Meta criada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meta'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    get:
      tags: [Metas]
      summary: Listar metas
      description: |
        Retorna todas as metas do usuário autenticado,
        ordenadas por `fechaLimite` **ascendente** (mais próximas de vencer primeiro).
        Atualmente não suporta filtros por query params.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de metas ordenadas por data limite ASC.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Meta'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /metas/{id}:
    get:
      tags: [Metas]
      summary: Obter meta por ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID da meta.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detalhe da meta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Metas]
      summary: Atualizar meta
      description: Atualiza os campos enviados. Retorna a meta atualizada.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID da meta.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMetaRequest'
      responses:
        '200':
          description: Meta atualizada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /metas/{id}/desativar:
    patch:
      tags: [Metas]
      summary: Desativar meta (soft delete)
      description: Marca a meta como inativa. Não retorna body.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID da meta.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Meta desativada. Sem body de resposta.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ════════════════════════════════════════
  #  ALERTAS
  # ════════════════════════════════════════
...